Add structure for separating positive and negative verb forms.
Verb conjugations will have to be separated, so the current approach is to sepearate based on polarity. Tense and formality will be parameters in the methods. This appraoch will help to decrease the logic for each method when determining polarity.

Refomat the get_verb_stem function to a more general splice_verb such that the function can handle return the verb stem or ending particles. This removes any obj[:int] code when retrieving the final particles throughout the code.